get 'mex.h' out of controlUtil.h
