+++ systems/controllers/individualCentersOfPressuremex.cpp
@@ -1,6 +1,7 @@
 #include "mex.h"
 #include "controlUtil.h"
 #include "drakeMexUtil.h"
+#include "controlMexUtil.h"
 #include <iostream>
 
 using namespace std;
